refactor: split data and view logic
