+++ lua/call_graph/init.lua
@@ -7,43 +7,15 @@ local M = {
     auto_toggle_hl = true
   }
 }
+local Caller = require("call_graph.caller")
 local log = require("call_graph.utils.log")
-local incoming_callers = {}
 
-local function get_len(t)
-  local len = 0
-  for _, _ in pairs(t) do
-    len = len + 1
-  end
-  return len
-end
 
 local function create_user_cmd()
   vim.api.nvim_create_user_command(
     "CallGraph",
     function()
-      local function gen_graph_done_cb(bufnr, ctx)
-        local caller = ctx
-        incoming_callers[tostring(bufnr)] = caller
-      end
-      local caller
-      if not M.opts.reuse_buf then
-        caller = require("call_graph.caller"):new(M.opts.hl_delay_ms, M.opts.auto_toggle_hl)
-      else
-        if get_len(incoming_callers) == 0 then
-          caller = require("call_graph.caller"):new(M.opts.hl_delay_ms, M.opts.auto_toggle_hl)
-        else
-          for _, c in pairs(incoming_callers) do
-            caller = c
-          end
-          caller = caller:reset_graph() -- reset graph will keep hl_delay_ms and auto_toggle_hl attr
-        end
-      end
-      caller.gen_graph_done = {
-        cb = gen_graph_done_cb,
-        cb_ctx = caller
-      }
-      caller:generate_call_graph()
+      Caller.generate_call_graph(M.opts)
     end,
     { desc = "Generate call graph for current buffer" }
   )
@@ -55,17 +27,13 @@ local function create_user_cmd()
       switch = "off"
     end
     vim.notify(string.format("Call graph reuse buf is %s", switch), vim.log.levels.INFO)
-  end, {desc = "Toggle reuse buf of call graph"})
+  end, { desc = "Toggle reuse buf of call graph" })
 
   vim.api.nvim_create_user_command("CallGraphToggleAutoHighlight", function()
     M.opts.auto_toggle_hl = not M.opts.auto_toggle_hl
     local switch = "on"
     if not M.opts.auto_toggle_hl then
       switch = "off"
-    end
-    vim.notify(string.format("Call graph auto highlighting is %s", switch), vim.log.levels.INFO)
-    for _, caller in pairs(incoming_callers) do
-      caller:set_auto_toggle_hl(M.opts.auto_toggle_hl)
     end
   end, { desc = "Toggle highlighting of call graph" })
 
@@ -76,22 +44,10 @@ local function create_user_cmd()
     , { desc = "Open call graph log file" })
 end
 
-local function buf_del_cb(bufnr)
-  log.debug("del buf", bufnr)
-  incoming_callers[tostring(bufnr)] = nil -- remove caller
-end
-
 local function setup_hl()
   vim.api.nvim_set_hl(0, "CallGraphLine", { link = "Search" })
 end
 
-local function setup_autocmd()
-  vim.api.nvim_create_autocmd({ "BufDelete", "BufWipeout" }, { -- 同时监听 BufDelete 和 BufWipeout 事件
-    callback = function(event)
-      buf_del_cb(event.buf)
-    end,
-  })
-end
 
 function M.setup(opts)
   -- setup opts
@@ -104,7 +60,6 @@ function M.setup(opts)
   log.setup({ append = false, level = M.opts.log_level })
   setup_hl()
   create_user_cmd()
-  setup_autocmd()
 end
 
 return M
